IECoreHoudini FnParameterisedHolder : Port to python bindings.
