--- include/IECoreHoudini/FnParameterisedHolder.h
@@ -1,81 +1,0 @@
-//////////////////////////////////////////////////////////////////////////
-//
-//  Copyright 2010 Dr D Studios Pty Limited (ACN 127 184 954) (Dr. D Studios),
-//  its affiliates and/or its licensors.
-//
-//  Copyright (c) 2010-2013, Image Engine Design Inc. All rights reserved.
-//
-//  Redistribution and use in source and binary forms, with or without
-//  modification, are permitted provided that the following conditions are
-//  met:
-//
-//     * Redistributions of source code must retain the above copyright
-//       notice, this list of conditions and the following disclaimer.
-//
-//     * Redistributions in binary form must reproduce the above copyright
-//       notice, this list of conditions and the following disclaimer in the
-//       documentation and/or other materials provided with the distribution.
-//
-//     * Neither the name of Image Engine Design nor the names of any
-//       other contributors to this software may be used to endorse or
-//       promote products derived from this software without specific prior
-//       written permission.
-//
-//  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
-//  IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
-//  THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
-//  PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR
-//  CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
-//  EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
-//  PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
-//  PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
-//  LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
-//  NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-//  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-//
-//////////////////////////////////////////////////////////////////////////
-
-#ifndef IECOREHOUDINI_FNPARAMETERISEDHOLDER_H
-#define IECOREHOUDINI_FNPARAMETERISEDHOLDER_H
-
-#include "IECoreHoudini/NodeHandle.h"
-#include "IECoreHoudini/ParameterisedHolderInterface.h"
-
-#include "OP/OP_Node.h"
-
-namespace IECoreHoudini
-{
-
-/// \todo: This class is a bit funny. Can this all be moved to the bindings like in IECoreMaya?
-class FnParameterisedHolder
-{
-	public :
-
-		FnParameterisedHolder( OP_Node *node=0 );
-		virtual ~FnParameterisedHolder();
-
-		bool hasParameterised();
-
-		void setParameterised( IECore::RunTimeTypedPtr p );
-		void setParameterised( const std::string &className, int classVerison, const std::string &seachPathEnvVar );
-
-		/// Sets the values of the parameters of the held Parameterised object
-		/// to reflect the values of the attributes of the node.
-		/// \todo: add setNodeValues as well
-		void setParameterisedValues( double time );
-
-		IECore::RunTimeTypedPtr getParameterised();
-
-	private :
-
-		bool hasHolder();
-		void setHolder( OP_Node *node );
-		ParameterisedHolderInterface *getHolder( OP_Node *node );
-
-		NodeHandle m_handle;
-
-};
-
-}
-
-#endif // IECOREHOUDINI_FNPARAMETERISEDHOLDER_H
